requirements.txt
absl-py==1.0.0
appnope==0.1.3
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
Bottleneck @ file:///opt/concourse/worker/volumes/live/51dc1780-97b5-4c4b-5601-ba5845532b9c/volume/bottleneck_1648028903302/work
cachetools==5.0.0
category-encoders==1.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.0
cornac==1.14.2
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy==1.6.0
decorator==5.1.1
deepctr==0.9.0
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
flatbuffers==2.0
fonttools==4.25.0
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.1
h5py==2.10.0
huggingface-hub==0.5.1
hypothesis==6.46.1
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
Jinja2==3.0.3
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jupyter-client==7.3.0
jupyter-core==4.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///Users/builder/miniconda3/envs/prefect/conda-bld/kiwisolver_1638569885739/work
libclang==14.0.1
lightfm==1.16
lightgbm @ file:///Users/runner/miniforge3/conda-bld/lightgbm_1641600026901/work
llvmlite==0.38.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_croot-hkd3cnd2/matplotlib-suite_1647506470899/work
matplotlib-inline==0.1.3
memory-profiler==0.60.0
mkl-fft==1.3.1
mkl-random @ file:///opt/concourse/worker/volumes/live/f196a661-8e33-4aac-63bc-efb2ff50e035/volume/mkl_random_1626186080069/work
mkl-service==2.4.0
mpmath==1.2.1
munkres==1.1.4
mypy-extensions==0.4.3
nest-asyncio==1.5.5
nltk==3.7
numba==0.55.1
numexpr @ file:///opt/concourse/worker/volumes/live/52f18ae0-3912-48f4-7832-7d22abf43616/volume/numexpr_1640704220628/work
numpy @ file:///opt/concourse/worker/volumes/live/29d88eb7-79e1-44fd-5f44-5ca283271b97/volume/numpy_and_numpy_base_1649782775897/work
oauthlib==3.2.0
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.2
pandera==0.11.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
powerlaw==1.5
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
Pygments==2.12.0
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
recommenders==1.1.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
retrying==1.3.3
rsa==4.8
sacremoses==0.0.50
scikit-learn @ file:///opt/concourse/worker/volumes/live/09b793c0-41c8-469b-7e31-b4a39bc68d95/volume/scikit-learn_1642617120730/work
scikit-surprise==1.1.1
scipy @ file:///opt/concourse/worker/volumes/live/7554f9b4-1616-4d3d-50f8-914545975614/volume/scipy_1641555024289/work
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn-pandas==2.2.0
sklearn2pmml==0.81.0
sortedcontainers==2.4.0
stack-data==0.2.0
statsmodels==0.13.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tokenizers==0.12.1
tornado @ file:///opt/concourse/worker/volumes/live/05341796-4198-4ded-4a9a-332fde3cdfd1/volume/tornado_1606942323372/work
tqdm @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_6cb3db64-b7e1-4db2-ab0b-f944920e5367lekntvc5/croots/recipe/tqdm_1650891088614/work
traitlets==5.1.1
transformers==4.18.0
typing-inspect==0.7.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
xgboost==1.5.0
zipp==3.8.0
